+++ castor/lib/project/Dependencies.scala
@@ -2,13 +2,13 @@ import sbt._
 
 object Dependencies {
   object Versions {
-    val zio = "2.0.2"
+    val zio = "2.0.4"
     val doobie = "1.0.0-RC2"
     val zioCatsInterop = "3.3.0"
-    val prismNodeClient = "0.1.0"
-    val prismSdk = "v1.4.1"
+    val prismNodeClient = "0.2.0"
+    val prismSdk = "v1.4.1" // scala-steward:off
     val shared = "0.2.0"
-    val flyway = "9.7.0"
+    val flyway = "9.8.3"
   }
 
   private lazy val zio = "dev.zio" %% "zio" % Versions.zio