+++ infrastructure/local/run.sh
@@ -9,12 +9,14 @@ Help()
    # Display Help
    echo "Run an instance of the ATALA bulding-block stack locally"
    echo
-   echo "Syntax: run.sh [-n/--name NAME|-p/--port PORT|-b/--background|-h/--help]"
+   echo "Syntax: run.sh [-n/--name NAME|-p/--port PORT|-b/--background|-e/--env|-w/--wait|-h/--help]"
    echo "options:"
    echo "-n/--name              Name of this instance - defaults to dev."
    echo "-p/--port              Port to run this instance on - defaults to 80."
    echo "-b/--background        Run in docker-compose daemon mode in the background."
    echo "-w/--wait              Wait until all containers are healthy (only in the background)."
+   echo "-e/--env               Provide your own .env file with versions."
+   echo "--debug                Run additional services for debug using docker-compose debug profile."
    echo "-h/--help              Print this help text."
    echo
 }
@@ -39,6 +41,15 @@ while [[ $# -gt 0 ]]; do
       ;;
     -w|--wait)
       WAIT="--wait"
+      shift # past argument
+      ;;
+    -e|--env)
+      ENV_FILE="$2"
+      shift # past argument
+      shift # past value
+      ;;
+    --debug)
+      DEBUG="--profile debug"
       shift # past argument
       ;;
     -h|--help)
@@ -66,15 +77,18 @@ fi
 
 NAME="${NAME:=local}"
 PORT="${PORT:=80}"
+ENV_FILE="${ENV_FILE:=${SCRIPT_DIR}/.env}"
+
 
 echo "NAME            = ${NAME}"
 echo "PORT            = ${PORT}"
+echo "ENV_FILE        = ${ENV_FILE}"
 
 echo "--------------------------------------"
-echo "Bringing up stack using docker-compose"
+echo "Starting stack using docker compose"
 echo "--------------------------------------"
 
-PORT=${PORT} docker-compose \
+PORT=${PORT} docker compose \
   -p ${NAME} \
   -f ${SCRIPT_DIR}/../shared/docker-compose.yml \
-  --env-file ${SCRIPT_DIR}/.env up ${BACKGROUND} ${WAIT}
+  --env-file ${ENV_FILE} ${DEBUG} up ${BACKGROUND} ${WAIT}