+++ infrastructure/local/stop.sh
@@ -60,10 +60,10 @@ PORT="${PORT:=80}"
 echo "NAME            = ${NAME}"
 
 echo "--------------------------------------"
-echo "Stopping stack using docker-compose"
+echo "Stopping stack using docker compose"
 echo "--------------------------------------"
 
-PORT=${PORT} docker-compose \
+PORT=${PORT} docker compose \
   -p ${NAME} \
   -f ${SCRIPT_DIR}/../shared/docker-compose.yml \
   --env-file ${SCRIPT_DIR}/.env down ${VOLUMES}