+++ prism-agent/service/project/Dependencies.scala
@@ -2,20 +2,20 @@ import sbt._
 
 object Dependencies {
   object Versions {
-    val zio = "2.0.2"
+    val zio = "2.0.4"
     val zioConfig = "3.0.2"
     val zioHttp = "2.0.0-RC11"
-    val zioInteropCats = "3.3.0"
+    val zioInteropCats = "3.3.0" // scala-steward:off
     val akka = "2.6.20"
     val akkaHttp = "10.2.9"
-    val castor = "0.4.0"
-    val pollux = "0.6.0-SNAPSHOT"
-    val connect = "0.3.0-SNAPSHOT"
+    val castor = "0.4.1"
+    val pollux = "0.8.0"
+    val connect = "0.3.1"
     val bouncyCastle = "1.70"
-    val logback = "1.4.4"
-    val mercury = "0.10.0"
+    val logback = "1.4.5"
+    val mercury = "0.10.1"
     val zioJson = "0.3.0"
-    val tapir = "1.2.2"
+    val tapir = "1.2.3"
   }
 
   private lazy val zio = "dev.zio" %% "zio" % Versions.zio
@@ -63,10 +63,10 @@ object Dependencies {
   private lazy val http4sBlazeServer = "org.http4s" %% "http4s-blaze-server" % "0.23.12"
 
   private lazy val tapirRedocBundle = "com.softwaremill.sttp.tapir" %% "tapir-redoc-bundle" % Versions.tapir
+
   private lazy val tapirSttpStubServer =
     "com.softwaremill.sttp.tapir" %% "tapir-sttp-stub-server" % Versions.tapir % Test
-
-  private lazy val sttpClient3ZioJson = "com.softwaremill.sttp.client3" %% "zio-json" % "3.8.0" % Test
+  private lazy val sttpClient3ZioJson = "com.softwaremill.sttp.client3" %% "zio-json" % "3.8.3" % Test
 
   // Dependency Modules
   private lazy val baseDependencies: Seq[ModuleID] = Seq(