+++ tests/e2e-tests/src/test/kotlin/features/did/ResolveDidSteps.kt
@@ -1,23 +1,23 @@
 package features.did
 
-import extentions.WithAgents
+import common.Agents.Acme
 import io.cucumber.java.en.Then
 import io.cucumber.java.en.When
 import net.serenitybdd.screenplay.rest.interactions.Get
 import net.serenitybdd.screenplay.rest.questions.ResponseConsequence
 
-class ResolveDidSteps: WithAgents() {
+class ResolveDidSteps {
 
     @When("I resolve existing DID by DID reference")
     fun iResolveExistingDIDByDIDReference() {
-        acme.attemptsTo(
+        Acme.attemptsTo(
             Get.resource("/connections")
         )
     }
 
     @Then("Response code is 200")
     fun responseCodeIs() {
-        acme.should(
+        Acme.should(
             ResponseConsequence.seeThatResponse("DID has required fields") {
                 it.statusCode(200)
             }