+++ tests/e2e-tests/src/test/kotlin/features/issue_credentials/IssueCredentialsSteps.kt
@@ -8,94 +8,31 @@ import net.serenitybdd.screenplay.rest.interactions.Post
 import net.serenitybdd.screenplay.rest.questions.ResponseConsequence
 import api_models.Connection
 import api_models.Credential
-import extentions.WithAgents
+import common.Agents.Acme
+import common.Agents.Bob
+import common.Utils.lastResponse
+import common.Utils.wait
+import features.connection.ConnectionSteps
 
-class IssueCredentialsSteps : WithAgents() {
+class IssueCredentialsSteps {
 
     @Given("Acme and Bob have an existing connection")
     fun acmeAndBobHaveAnExistingConnection() {
-        // Acme(Issuer) initiates a connection
-        // and sends it to Bob(Holder) out-of-band, e.g. using QR-code
-        acme.attemptsTo(
-            Post.to("/connections")
-                .with {
-                    it.header("Content-Type", "application/json")
-                    it.body("""{"label": "Connect with ${bob.name}"}""")
-                }
-        )
-        acme.should(
-            ResponseConsequence.seeThatResponse("Creates connection request") {
-                it.statusCode(201)
-            }
-        )
-        val acmeConnection = lastResponse().getObject("", Connection::class.java)
-
-        // Here out of band transfer of connection QR code is happening
-        // and Bob (Holder) gets an invitation URL
-        // they're accepting connection invitation by POST request specifying achieved invitation
-        bob.remember("invitationUrl", acmeConnection.invitation.invitationUrl.split("=")[1])
-
-        // Bob accepts connection using achieved out-of-band invitation
-        bob.attemptsTo(
-            Post.to("/connection-invitations")
-                .with {
-                    it.header("Content-Type", "application/json")
-                    it.body("""{"invitation": "${bob.recall<String>("invitationUrl")}"}""")
-                }
-        )
-        bob.should(
-            ResponseConsequence.seeThatResponse("Accepts connection request") {
-                it.statusCode(200)
-            }
-        )
-        val bobConnection = lastResponse().getObject("", Connection::class.java)
-
-        // Acme(Issuer) checks their connections to check if invitation was accepted by Bob(Holder)
-        // and sends final connection response
-        wait(
-            {
-                acme.attemptsTo(
-                    Get.resource("/connections/${acmeConnection.connectionId}"),
-                )
-                acme.should(
-                    ResponseConsequence.seeThatResponse("Get issuer connections") {
-                        it.statusCode(200)
-                    }
-                )
-                lastResponse()
-                    .getObject("", Connection::class.java).state == "ConnectionResponseSent"
-            },
-            "Issuer did not sent final connection confirmation! Connection didn't reach ConnectionResponseSent state."
-        )
-        acme.remember("did", lastResponse().getObject("", Connection::class.java).myDid)
-        acme.remember("holderDid", lastResponse().getObject("", Connection::class.java).theirDid)
-
-        // Bob (Holder) receives final connection response
-        wait(
-            {
-                bob.attemptsTo(
-                    Get.resource("/connections/${bobConnection.connectionId}")
-                )
-                bob.should(
-                    ResponseConsequence.seeThatResponse("Get holder connections") {
-                        it.statusCode(200)
-                    }
-                )
-                lastResponse().getObject("", Connection::class.java).state == "ConnectionResponseReceived"
-            },
-            "Holder did not receive final connection confirmation! Connection didn't reach ConnectionResponseReceived state."
-        )
-
-        bob.remember("did", lastResponse().getObject("", Connection::class.java).myDid)
-        bob.remember("issuerDid", lastResponse().getObject("", Connection::class.java).theirDid)
-        // Connection established. Both parties exchanged their DIDs with each other
+        val connectionSteps = ConnectionSteps()
+        connectionSteps.inviterGeneratesAConnectionInvitation()
+        connectionSteps.inviteeReceivesTheConnectionInvitation()
+        connectionSteps.inviteeSendsAConnectionRequestToInviter()
+        connectionSteps.inviterReceivesTheConnectionRequest()
+        connectionSteps.inviterSendsAConnectionResponseToInvitee()
+        connectionSteps.inviteeReceivesTheConnectionResponse()
+        connectionSteps.inviterAndInviteeHaveAConnection()
     }
 
     @Given("Acme offers a credential")
     fun acmeOffersACredential() {
         val newCredential = Credential(
             schemaId = "schema:1234",
-            subjectId = acme.recall("holderDid"),
+            subjectId = Acme.recall<Connection>("connection").theirDid,
             validityPeriod = 3600,
             automaticIssuance = false,
             awaitConfirmation = false,
@@ -105,14 +42,14 @@ class IssueCredentialsSteps : WithAgents() {
             )
         )
 
-        acme.attemptsTo(
+        Acme.attemptsTo(
             Post.to("/issue-credentials/credential-offers")
                 .with {
                     it.header("Content-Type", "application/json")
                     it.body(newCredential)
                 }
         )
-        acme.should(
+        Acme.should(
             ResponseConsequence.seeThatResponse("The issue credential offer created") {
                 it.statusCode(201)
             }
@@ -125,10 +62,10 @@ class IssueCredentialsSteps : WithAgents() {
     fun bobRequestsTheCredential() {
         wait(
             {
-                bob.attemptsTo(
+                Bob.attemptsTo(
                     Get.resource("/issue-credentials/records")
                 )
-                bob.should(
+                Bob.should(
                     ResponseConsequence.seeThatResponse("Credential records") {
                         it.statusCode(200)
                     }
@@ -140,12 +77,12 @@ class IssueCredentialsSteps : WithAgents() {
 
         val recordId = lastResponse().getList("items", Credential::class.java)
             .findLast { it.protocolState == "OfferReceived" }!!.recordId
-        bob.remember("recordId", recordId)
+        Bob.remember("recordId", recordId)
 
-        bob.attemptsTo(
+        Bob.attemptsTo(
             Post.to("/issue-credentials/records/${recordId}/accept-offer")
         )
-        bob.should(
+        Bob.should(
             ResponseConsequence.seeThatResponse("Accept offer") {
                 it.statusCode(200)
             }
@@ -156,10 +93,10 @@ class IssueCredentialsSteps : WithAgents() {
     fun acmeIssuesTheCredential() {
         wait(
             {
-                acme.attemptsTo(
+                Acme.attemptsTo(
                     Get.resource("/issue-credentials/records")
                 )
-                acme.should(
+                Acme.should(
                     ResponseConsequence.seeThatResponse("Credential records") {
                         it.statusCode(200)
                     }
@@ -171,10 +108,10 @@ class IssueCredentialsSteps : WithAgents() {
         )
         val recordId = lastResponse().getList("items", Credential::class.java)
             .findLast { it.protocolState == "RequestReceived" }!!.recordId
-        acme.attemptsTo(
+        Acme.attemptsTo(
             Post.to("/issue-credentials/records/${recordId}/issue-credential")
         )
-        acme.should(
+        Acme.should(
             ResponseConsequence.seeThatResponse("Issue credential") {
                 it.statusCode(200)
             }
@@ -182,10 +119,10 @@ class IssueCredentialsSteps : WithAgents() {
 
         wait(
             {
-                acme.attemptsTo(
+                Acme.attemptsTo(
                     Get.resource("/issue-credentials/records/${recordId}")
                 )
-                acme.should(
+                Acme.should(
                     ResponseConsequence.seeThatResponse("Credential records") {
                         it.statusCode(200)
                     }
@@ -200,10 +137,10 @@ class IssueCredentialsSteps : WithAgents() {
     fun bobHasTheCredentialIssued() {
         wait(
             {
-                bob.attemptsTo(
-                    Get.resource("/issue-credentials/records/${bob.recall<String>("recordId")}")
+                Bob.attemptsTo(
+                    Get.resource("/issue-credentials/records/${Bob.recall<String>("recordId")}")
                 )
-                bob.should(
+                Bob.should(
                     ResponseConsequence.seeThatResponse("Credential records") {
                         it.statusCode(200)
                     }